+++ mapjs/src/npm-main.js
@@ -3,7 +3,6 @@
 require('./browser/dom-map-widget');
 require('./browser/map-toolbar-widget');
 require('./browser/link-edit-widget');
-require('./argmapjs-utilities.js');
 require('./browser/image-drop-widget');
 
 module.exports = {